requirements.txt
# Dev environment
pip
autopep8

# App
flask
openai
python-dotenv